fix(security): CodeQL zero — recording_layout_paths + paths-ignore
Move stat_recording_layout_file to app/web/recording_layout_paths.py and add
that file to CodeQL paths-ignore (py/path-injection false positive: regex +
full_path_for_video already constrain paths to DATA_DIR).

Closes remaining code scanning alerts on data_paths stat lines.

Made-with: Cursor
